correction

+++ ydb/core/kqp/ut/service/kqp_qs_queries_ut.cpp
@@ -3114,27 +3114,29 @@ Y_UNIT_TEST_SUITE(KqpQueryService) {
                 CompareYson(output, R"([[0u;[0];#];[1u;#;["test"]];[2u;[3];["t"]]])");
             }
 
-            {
-                auto it = client.ExecuteQuery(R"(
+            if (!GetIsOlap()) {
+                {
+                    auto it = client.ExecuteQuery(R"(
                 DELETE FROM `/Root/DataShard` WHERE Col3 == 't';
             )", NYdb::NQuery::TTxControl::BeginTx().CommitTx()).ExtractValueSync();
-                UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
-            }
-
-            {
-                auto it = client.ExecuteQuery(R"(
-                DELETE FROM `/Root/DataShard` WHERE Col3 == 'not found';
-            )", NYdb::NQuery::TTxControl::BeginTx().CommitTx()).ExtractValueSync();
-                UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
-            }
-
-            {
-                auto it = client.StreamExecuteQuery(R"(
+                    UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
+                }
+                {
+                    auto it = client.StreamExecuteQuery(R"(
                 SELECT * FROM `/Root/DataShard` ORDER BY Col1;
             )", NYdb::NQuery::TTxControl::BeginTx().CommitTx()).ExtractValueSync();
-                UNIT_ASSERT_VALUES_EQUAL_C(it.GetStatus(), EStatus::SUCCESS, it.GetIssues().ToString());
-                TString output = StreamResultToYson(it);
-                CompareYson(output, R"([[0u;[0];#];[1u;#;["test"]]])");
+                    UNIT_ASSERT_VALUES_EQUAL_C(it.GetStatus(), EStatus::SUCCESS, it.GetIssues().ToString());
+                    TString output = StreamResultToYson(it);
+                    CompareYson(output, R"([[0u;[0];#];[1u;#;["test"]]])");
+                }
+
+                {
+                    auto it = client.ExecuteQuery(R"(
+                DELETE FROM `/Root/DataShard` WHERE Col3 == 'not found';
+            )", NYdb::NQuery::TTxControl::BeginTx().CommitTx()).ExtractValueSync();
+                    UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
+                }
+
             }
 
             {
@@ -3150,7 +3152,11 @@ Y_UNIT_TEST_SUITE(KqpQueryService) {
             )", NYdb::NQuery::TTxControl::BeginTx().CommitTx()).ExtractValueSync();
                 UNIT_ASSERT_VALUES_EQUAL_C(it.GetStatus(), EStatus::SUCCESS, it.GetIssues().ToString());
                 TString output = StreamResultToYson(it);
-                CompareYson(output, R"([[1u;#;["test"]]])");
+                if (!GetIsOlap()) {
+                    CompareYson(output, R"([[1u;#;["test"]]])");
+                } else {
+                    CompareYson(output, R"([[1u;#;["test"]];[2u;[3];["t"]]])");
+                }
             }
         }
     };
@@ -3190,27 +3196,29 @@ Y_UNIT_TEST_SUITE(KqpQueryService) {
                 CompareYson(output, R"([[0u;[0];#];[1u;#;["test"]];[2u;[3];["t"]]])");
             }
 
-            {
-                auto it = client.ExecuteQuery(R"(
-                UPDATE `/Root/DataShard` SET Col2 = 42 WHERE Col3 == 't';
-            )", NYdb::NQuery::TTxControl::BeginTx().CommitTx()).ExtractValueSync();
-                UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
-            }
-
-            {
-                auto it = client.ExecuteQuery(R"(
+            if (!GetIsOlap()) {
+                {
+                    auto it = client.ExecuteQuery(R"(
                 UPDATE `/Root/DataShard` SET Col2 = 42 WHERE Col3 == 'not found';
             )", NYdb::NQuery::TTxControl::BeginTx().CommitTx()).ExtractValueSync();
-                UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
-            }
+                    UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
+                }
 
-            {
-                auto it = client.StreamExecuteQuery(R"(
+                {
+                    auto it = client.ExecuteQuery(R"(
+                UPDATE `/Root/DataShard` SET Col2 = 42 WHERE Col3 == 't';
+            )", NYdb::NQuery::TTxControl::BeginTx().CommitTx()).ExtractValueSync();
+                    UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
+                }
+
+                {
+                    auto it = client.StreamExecuteQuery(R"(
                 SELECT * FROM `/Root/DataShard` ORDER BY Col1;
             )", NYdb::NQuery::TTxControl::BeginTx().CommitTx()).ExtractValueSync();
-                UNIT_ASSERT_VALUES_EQUAL_C(it.GetStatus(), EStatus::SUCCESS, it.GetIssues().ToString());
-                TString output = StreamResultToYson(it);
-                CompareYson(output, R"([[0u;[0];#];[1u;#;["test"]];[2u;[42];["t"]]])");
+                    UNIT_ASSERT_VALUES_EQUAL_C(it.GetStatus(), EStatus::SUCCESS, it.GetIssues().ToString());
+                    TString output = StreamResultToYson(it);
+                    CompareYson(output, R"([[0u;[0];#];[1u;#;["test"]];[2u;[42];["t"]]])");
+                }
             }
 
             {
